+++ app/commands/__init__.py
@@ -1,6 +1,5 @@
 from app.commands.buildfeed import BuildFeed
 from app.commands.migrateshapes import MigrateShapes
 from app.commands.exportcsv import ExportCSV
-from app.commands.dumpdata import DumpData
-from app.commands.dumpdata import LoadData
+from app.commands.dumpload import DumpData, LoadData
 from app.commands.deploy import Deploy
